- Create a user context for a user specific logic - Refactor desktop user dropdown to get menu items from user context - Add  credentials provider component to encapsulate credentials logic there
