--- src/context/credentialsContext.js
@@ -1,6 +1,0 @@
-import { createContext } from "react";
-
-export const CredentialsContext = createContext({
-  storedCredentials: {},
-  setStoredCredentials: () => {},
-});
